+++ api/v1alpha1/obcluster_webhook.go
@@ -147,6 +147,14 @@ func (r *OBCluster) ValidateUpdate(old runtime.Object) (admission.Warnings, erro
 	} else if oldMode != oceanbaseconst.ModeStandalone && (oldCluster.Spec.OBServerTemplate.Resource.Cpu != r.Spec.OBServerTemplate.Resource.Cpu || oldCluster.Spec.OBServerTemplate.Resource.Memory != r.Spec.OBServerTemplate.Resource.Memory) {
 		return nil, errors.New("forbid to modify cpu or memory quota of non-standalone cluster")
 	}
+	if r.Spec.BackupVolume == nil && oldCluster.Spec.BackupVolume != nil {
+		return nil, errors.New("forbid to remove backup volume")
+	}
+	if r.Spec.BackupVolume != nil && oldCluster.Spec.BackupVolume == nil &&
+		mode != oceanbaseconst.ModeStandalone &&
+		mode != oceanbaseconst.ModeService {
+		return nil, errors.New("forbid to add backup volume on non-service or non-standalone cluster")
+	}
 	var err error
 	if r.Spec.OBServerTemplate.Storage.DataStorage.Size.Cmp(oldCluster.Spec.OBServerTemplate.Storage.DataStorage.Size) > 0 {
 		err = errors.Join(err, r.validateStorageClassAllowExpansion(r.Spec.OBServerTemplate.Storage.DataStorage.StorageClass))
