+++ internal/const/status/obcluster/obcluster_status.go
@@ -25,4 +25,5 @@ const (
 	ScaleUp             = "scale up"
 	ExpandPVC           = "expand pvc"
 	Failed              = "failed"
+	MountBackupVolume   = "mount backup volume"
 )
