+++ internal/const/status/observer/observer_status.go
@@ -25,4 +25,5 @@ const (
 	ScaleUp           = "scale up"
 	ExpandPVC         = "expand pvc"
 	FinalizerFinished = "finalizer finished"
+	MountBackupVolume = "mount backup volume"
 )
