+++ internal/resource/observer/observer_flow.go
@@ -126,3 +126,16 @@ func ResizePVC() *tasktypes.TaskFlow {
 		},
 	}
 }
+
+func MountBackupVolume() *tasktypes.TaskFlow {
+	return &tasktypes.TaskFlow{
+		OperationContext: &tasktypes.OperationContext{
+			Name:         fMountBackupVolume,
+			Tasks:        []tasktypes.TaskName{tDeletePod, tWaitForPodDeleted, tCreateOBPod, tWaitOBServerReady},
+			TargetStatus: serverstatus.Running,
+			OnFailure: tasktypes.FailureRule{
+				Strategy: strategy.StartOver,
+			},
+		},
+	}
+}
