+++ internal/resource/obzone/init.go
@@ -28,4 +28,5 @@ func init() {
 	task.GetRegistry().Register(fDeleteOBZoneFinalizer, DeleteOBZoneFinalizer)
 	task.GetRegistry().Register(fScaleUpOBServers, ScaleUpOBServers)
 	task.GetRegistry().Register(fExpandPVC, ResizePVC)
+	task.GetRegistry().Register(fMountBackupVolume, MountBackupVolume)
 }
