+++ internal/resource/obzone/obzone_flow.go
@@ -120,3 +120,13 @@ func ResizePVC() *tasktypes.TaskFlow {
 		},
 	}
 }
+
+func MountBackupVolume() *tasktypes.TaskFlow {
+	return &tasktypes.TaskFlow{
+		OperationContext: &tasktypes.OperationContext{
+			Name:         fMountBackupVolume,
+			Tasks:        []tasktypes.TaskName{tMountBackupVolume, tWaitForOBServerMounting, tWaitOBServerRunning},
+			TargetStatus: zonestatus.Running,
+		},
+	}
+}
